💄 Make select list on macOS a bit smaller
It caused the lower end of the select lists to overlap with the buttons
below.

I realize that this apparently was added specifically for macOS but I
can’t find any justification for it, and it breaks the layout, so
reverting that particular change seems like the prudent way to fix this.